requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
Deprecated==1.2.14
Django==3.2.1
django-cors-headers==3.10.0
django-crispy-forms==2.0
django-environ==0.10.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
redis==4.3.3
requests==2.28.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
vine==1.3.0
wrapt==1.15.0
